Commonize page transition animation
