+++ animations/page_transitions/fadeout1.js
@@ -1,0 +1,5 @@
+export default function fadeout1(selector) {
+  requestAnimationFrame(() => {
+    TweenMax.to(selector, 1, {opacity: 0});
+  });
+};
